ISSUE #2901 refactor validateNewRevisionData for both fed and containers
