+++ backend/src/v5/middleware/dataConverter/inputs/teamspaces/projects/models/containers.js
@@ -18,6 +18,7 @@
 const { createResponseCode, templates } = require('../../../../../../utils/responseCodes');
 const { getModelByQuery } = require('../../../../../../models/modelSettings');
 const { respond } = require('../../../../../../utils/responder');
+const { validateNewRevisionData } = require('./commons/revisions');
 
 const Containers = {};
 
@@ -34,5 +35,7 @@ Containers.canDeleteContainer = async (req, res, next) => {
 		respond(req, res, createResponseCode(templates.invalidArguments, err?.message));
 	}
 };
+
+Containers.validateNewRevisionData = validateNewRevisionData(false);
 
 module.exports = Containers;
