+++ backend/tests/v5/unit/middleware/dataConverter/inputs/teamspaces/projects/models/containers.test.js
@@ -15,7 +15,14 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-const { src } = require('../../../../../../../helper/path');
+const { src, modelFolder } = require('../../../../../../../helper/path');
+const MockExpressRequest = require('mock-express-request');
+const FormData = require('form-data');
+const fs = require('fs');
+const path = require('path');
+
+jest.mock('../../../../../../../../../src/v5/utils/quota');
+const Quota = require(`${src}/utils/quota`);
 
 jest.mock('../../../../../../../../../src/v5/utils/responder');
 const Responder = require(`${src}/utils/responder`);
@@ -57,6 +64,58 @@ const testCanDeleteContainer = () => {
 	});
 };
 
+const createRequestWithFile = (teamspace, { tag, desc, importAnim }, unsupportedFile = false, noFile = false) => {
+	const form = new FormData();
+	if (!noFile) {
+		form.append('file',
+			fs.createReadStream(path.join(modelFolder, unsupportedFile ? 'dummy.png' : 'dummy.obj')));
+	}
+	if (tag) form.append('tag', tag);
+	if (desc) form.append('desc', desc);
+	if (importAnim) form.append('importAnim', importAnim);
+
+	const req = new MockExpressRequest({
+		method: 'POST',
+		host: 'localhost',
+		url: `/${teamspace}/upload`,
+		headers: form.getHeaders(),
+	});
+
+	form.pipe(req);
+	req.params = { teamspace };
+	return req;
+};
+
+Quota.sufficientQuota.mockImplementation((ts) => (ts === 'noQuota' ? Promise.reject(templates.quotaLimitExceeded) : Promise.resolve()));
+
+const testValidateNewRevisionData = () => {
+	const standardBody = { tag: '123', description: 'this is a model', importAnim: false };
+	describe.each([
+		['Request with valid data', 'ts', standardBody],
+		['Request with unsupported model file', 'ts', standardBody, true, false, templates.unsupportedFileFormat],
+		['Request with insufficient quota', 'noQuota', standardBody, false, false, templates.quotaLimitExceeded],
+		['Request with no body should fail', 'ts', {}, false, false, templates.invalidArguments],
+		['Request with just tag should pass', 'ts', { tag: 'dkf_j-d' }, false, false],
+		['Request with wrong tag type should fail', 'ts', { tag: false }, false, false, templates.invalidArguments],
+		['Request with tag that is not alphanumeric should fail', 'ts', { tag: '1%2%3' }, false, false, templates.invalidArguments],
+		['Request with no file should fail', 'ts', { tag: 'drflgdf' }, false, true, templates.invalidArguments],
+	])('Check new revision data', (desc, ts, bodyContent, badFile, noFile, error) => {
+		test(`${desc} should ${error ? `fail with ${error.code}` : ' succeed and next() should be called'}`, async () => {
+			const req = createRequestWithFile(ts, bodyContent, badFile, noFile);
+			const mockCB = jest.fn(() => {});
+			await Containers.validateNewRevisionData(req, {}, mockCB);
+			if (error) {
+				expect(mockCB.mock.calls.length).toBe(0);
+				expect(Responder.respond.mock.calls.length).toBe(1);
+				expect(Responder.respond.mock.results[0].value.code).toEqual(error.code);
+			} else {
+				expect(mockCB.mock.calls.length).toBe(1);
+			}
+		});
+	});
+};
+
 describe('middleware/dataConverter/inputs/teamspaces/projects/models/containers', () => {
 	testCanDeleteContainer();
+	testValidateNewRevisionData();
 });
